Changes to the glyph drawing API and fonts.
With this change, the font is no longer part of the graphics context
state. The mt_gc_set_font() API has been removed. Now, the font is
defined by the mt_item that's passed to mt_gc_draw_glyphs(). The
rationale for this is that it simplifies font management with respect
to font fallback.